Refactor ingestion/publish logic
 - Separate compatible versions of ingestion/publish
 - Extract common functions of ingestion/publish
 - Remove storage moduel gen_key function.

Closes: #157
